Update the Roslyn solution to build our executables as both 32 and 64-bit. ***NO_CI***  (changeset 1384854)

+++ Src/Compilers/Core/NativeClient/native_client.cpp
@@ -132,7 +132,8 @@ bool GetExpectedProcessPath(
 {
     processPath.clear();
     processPath.resize(MAX_PATH);
-    while (GetModuleFileNameW(NULL, &processPath[0], processPath.size()) == 0)
+    while (GetModuleFileNameW(NULL, &processPath[0],
+		static_cast<DWORD>(processPath.size())) == 0)
     {
         if (GetLastError() == ERROR_INSUFFICIENT_BUFFER) 
         {
